# perf: use more minimal requests for getting call data