--- a/apps/frontend/src/graphql/call/getCall.graphql
+++ b/apps/frontend/src/graphql/call/getCall.graphql
@@ -1,5 +1,0 @@
-query getCall($callId: Int!) {
-  call(callId: $callId) {
-    ...call
-  }
-}