--- a/apps/frontend/src/graphql/call/getCallCloseDetails.graphql
+++ b/apps/frontend/src/graphql/call/getCallCloseDetails.graphql
@@ -1,0 +1,10 @@
+query getCallCloseDetails($callId: Int!) {
+  call(callId: $callId) {
+    isActive
+    isActiveInternal
+    referenceNumberFormat
+    endCallInternal
+    startCall
+    endCall
+  }
+}