--- a/apps/frontend/src/graphql/call/getCallSubmissionDetails.graphql
+++ b/apps/frontend/src/graphql/call/getCallSubmissionDetails.graphql
@@ -1,0 +1,6 @@
+query getCallSubmissionDetails($callId: Int!) {
+  call(callId: $callId) {
+    proposalWorkflowId
+    submissionMessage
+  }
+}