Closes: https://github.com/UserOfficeProject/issue-tracker/issues/1246

## Description
This PR optimizes the retrieval of call data to use more minimal requests.

Testing against my local we have some speed-ups on these. This is probably as we are not getting user data from the instruments object that was on the call object. 

| getCall | getCallSubmissionDetails | getCallCloseDetails |
| ------------- | ------------- | ------------- |
| 3.772686s | 0.049143s | 0.055274s |
| 2.950422s | 0.013006s | 0.016095s |
| 3.277789s | 0.013949s | 0.013914s |
| 2.960995s | 0.019828s | 0.014190s |
| 2.819570s | 0.011753s | 0.014440s |
| 2.835794s | 0.012031s | 0.013006s |
| 2.854073s | 0.012009s | 0.020088s |
| 2.717915s | 0.014634s | 0.019050s |
| 2.729184s | 0.015949s | 0.012913s |
| 2.710989s | 0.012208s | 0.019772s |


## Motivation and Context
The previous method of fetching call data was excessively extensive and had a negative impact on performance. This PR addresses this by using more minimal requests, which improves overall application performance and responsiveness.

## Changes
1. The `useCallData` hook has been replaced with `useCallCloseData` in `ProposalCreate.tsx`. This allows us to fetch only the necessary details for closing a call.
2. The `getCall` API method has been replaced with `getCallSubmissionDetails` in `ProposalSummary.tsx`, fetching only the necessary details for call submissions.
3. The `getCall` GraphQL query has been removed and replaced with two new queries: `getCallCloseDetails` and `getCallSubmissionDetails`. These new queries fetch only the necessary details for closing a call and for call submissions, respectively.
4. The `useCallData` hook in `useCallCloseData.ts` has been updated to use the new `getCallCloseDetails` query.

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->

## Fixes Jira Issue

[https://jira.esss.lu.se/browse/](https://jira.esss.lu.se/browse/)

## Depends On

<!--- Does this PR depend on another PR that should be merged first or at the same time -->

## Tests included/Docs Updated?

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->

- [ ] I have added tests to cover my changes.
- [ ] All relevant doc has been updated